Update Route53Pricing.md
Adding a direct link to the pricing for domain registration, as it is currently a real hassle to access the prices for registering a domain name with Route 53 with first having to go to the general pricing page, then having to search and scroll a long time to finally reach the particular paragraph. As Amazon is known to be the most customer-friendly company in the world, also the pricing for domain registration on AWS should be accessible in the most hassle-free way. :)